requirements.txt
streamlit>=1.41,<2.0
pandas>=2.2,<3.0
plotly-express>=0.4.1,<0.5
